feat: add fallback to archive on https failure
